+++ mocha.start.js
@@ -2,13 +2,16 @@
 'use strict';
 
 var assert = require('chai').assert;
+assert.equalObjects = function (a, b, m) {
+  assert.deepEqual(JSON.parse(JSON.stringify(a)), JSON.parse(JSON.stringify(b)), m || 'Objects should be equal!');
+};
 var mocha = require('mocha');
 var sinon = require('sinon');
 var DSRedisAdapter = require('./');
 var JSData = require('js-data');
 JSData.DSUtils.Promise = require('bluebird');
 
-var adapter, store, DSUtils, DSErrors, User, Post, Comment;
+var adapter, store, DSUtils, DSErrors, Profile, User, Post, Comment;
 
 var globals = module.exports = {
   fail: function (msg) {
@@ -56,17 +59,22 @@ beforeEach(function () {
   adapter = new DSRedisAdapter();
   DSUtils = JSData.DSUtils;
   DSErrors = JSData.DSErrors;
+  globals.Profile = global.Profile = Profile = store.defineResource({
+    name: 'profile'
+  });
   globals.User = global.User = User = store.defineResource({
     name: 'user',
     relations: {
       hasMany: {
         post: {
           localField: 'posts',
-          foreignKey: 'userId'
-        },
-        comment: {
-          localField: 'comments',
-          foreignKey: 'userId'
+          foreignKey: 'post'
+        }
+      },
+      hasOne: {
+        profile: {
+          localField: 'profile',
+          localKey: 'profileId'
         }
       }
     }
@@ -122,6 +130,8 @@ afterEach(function (done) {
   global.adapter = null;
 
   adapter.destroyAll(User).then(function () {
+    return adapter.destroyAll(Profile);
+  }).then(function () {
     return adapter.destroyAll(Post);
   }).then(function () {
     return adapter.destroyAll(Comment);